Fix functionality when block has multiple kv-pairs
When a source block is has keywords such as #+NAME: name, the :begin of
the block is located at the first keyword. This causes
`org-babel-tangle-single-block` to determine a wrong line number for the
start of the source block contents, so that it is different from when
using `org-babel-tangle-collect-blocks` to determine the first line,
causing the failure in the `org-src-context--edit-src-ad` its logic.

This commit fixes it by making sure that the cursor is on the line
starting with #+BEGIN_SRC line when the first line of contents gets
determined via ``org-babel-tangle-single-block`.